Move parser config and result models to parsers app
