+++ solution/backend/parsers/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ParsersConfig(AppConfig):
+    name = "parsers"
+    verbose_name = "Parser configuration and logging"
